+++ Libs/LibClassicDurations/CHANGELOG.md
@@ -1,6 +1,8 @@
 # LibClassicDurations
 
-## [1.61](https://github.com/rgd87/LibClassicDurations/tree/1.61) (2020-06-29)
-[Full Changelog](https://github.com/rgd87/LibClassicDurations/compare/1.60...1.61) [Previous Releases](https://github.com/rgd87/LibClassicDurations/releases)
+## [1.64](https://github.com/rgd87/LibClassicDurations/tree/1.64) (2020-08-06)
+[Full Changelog](https://github.com/rgd87/LibClassicDurations/compare/1.63...1.64) [Previous Releases](https://github.com/rgd87/LibClassicDurations/releases)
 
-- Purging old guids before merge  
+- Switched to Github Actions  
+- Debug messages for Ignite  
+- Bugfix